vhost-user-blk: add vhost-user-blk-pci host device to Qemu
Add a new vhost-user device type "vhost-user-blk-pci" to
accelerate virtio-blk processing in user space. Qemu uses a
char device to communicate with the I/O processing backend.

The feature is enabled by default with Linux, and depends on
CONFIG_VIRTIO configuration. To use this feature users can
start the qemu with the following command line like this:

qemu-system-x86_64 -m 1024 -object memory-backend-file, \
id=mem,size=1G, mem-path=/dev/hugepages,share=on \
-numa node,memdev=mem -hda /root/ubuntu.img \
-chardev socket,id=char0,path=/path/vhost.0 \
-net user,hostfwd=tcp::10022-:22 -net nic \
-device vhost-user-blk-pci,chardev=char0,\
logical_block_size=512,capacity=262144 \
-vnc 0.0.0.0:1 --enable-kvm

For 'capacity' parameter, this stands for the number of sectors
for this device, the sector is counted by 512 Bytes.

Change-Id: I18b7ee930d07c6d9cb051a50398965a37a1c402b
Signed-off-by: Changpeng Liu <changpeng.liu@intel.com>