virtio: don't check queue consitence in virtio_load()
QEMU keep track of avail and used ring position in cache
(vq->vring.cache). For virtio it is working fine but for vhost-user
those caches are not updated since all communication bypass QEMU RW
memory handers and those caches are not updated.

For now just disable the check but final fix should address this problem
in some other way.

Change-Id: I19cf5a818f338778ea28c462f4d2dbd01133e9b1
Signed-off-by: Pawel Wodkowski <pawelx.wodkowski@intel.com>