virtio: don't check queue consitence in virtio_load()
QEMU keep track of avail and used ring position in cache
(vq->vring.cache). For virtio it is working fine but for vhost-user
those caches are not updated since all communication bypass QEMU RW
memory handers and those caches are not updated.

For now just disable the check but final fix should address this problem
in some other way.

Change-Id: I31ef43ee64c8db60600743b11dcc6ef023f876c6
Signed-off-by: Pawel Wodkowski <pawelx.wodkowski@intel.com>