block/NVMe: Adding metadata support for non-lnvm device
This patch adds the ability to issue read/writes with
separate metadata buffer to NVMe devices using lba format
with non-zero ms field. The support is added for single
namespace devices only.

The backing file should be added on a per controller basis.
Lightnvm device is using the same file for storing md and
chunk state.

Signed-off-by: Kozlowski Mateusz <mateusz.kozlowski@intel.com>
Change-Id: I1b35386e669366caa756b616162e0d064ebb4186