block/NVMe: Adding metadata support for non-lnvm device
This patch adds the ability to issue read/writes with
separate metadata buffer to NVMe devices using lba format
with non-zero ms field. The support is added for single
namespace devices only.

The backing file should be added on a per controller basis.
Lightnvm device is using the same file for storing md and
chunk state.

Signed-off-by: Kozlowski Mateusz <mateusz.kozlowski@intel.com>
Change-Id: I1b35386e669366caa756b616162e0d064ebb4186
Reviewed-on: https://review.gerrithub.io/c/spdk/qemu/+/454270
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Wojciech Malikowski <wojciech.malikowski@intel.com>
Reviewed-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Tested-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>