block/NVMe: Adding metadata support for non-lnvm device
This patch adds the ability to issue read/writes with
separate metadata buffer to NVMe devices using lba format
with non-zero ms field. The support is added for single
namespace devices only.

The backing file should be added on a per controller basis.
Lightnvm device is using the same file for storing md and
chunk state.

Signed-off-by: Wojciech Malikowski <wojciech.malikowski@intel.com>
Change-Id: I49287d85358bc7b81d2612c09902de54f86d193f
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/qemu/+/2274
Tested-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>