block/NVMe: lnvm commands can have <4k buffer alignment
Changed handling of internal scatter gather lists when dealing
with an array of offsets (used in lnvm write/read commands only)
so they are properly split and handled when number of sgl entries
is greater than number of entries in offset_list.

This allows for more lax buffer alignment requirements, as previously
misaligned buffers would result in parts of the request to be sent to
offset 0 of the underlying device, corrupting data.

Signed-off-by: Kozlowski Mateusz <mateusz.kozlowski@intel.com>
Change-Id: Ic40e40507c2b590e5d71f4cecee4ea0662ad1ccf