block/NVMe: lnvm commands can have <4k buffer alignment
Changed handling of internal scatter gather lists when dealing
with an array of offsets (used in lnvm write/read commands only)
so they are properly split and handled when number of sgl entries
is greater than number of entries in offset_list.

This allows for more lax buffer alignment requirements, as previously
misaligned buffers would result in parts of the request to be sent to
offset 0 of the underlying device, corrupting data.

Signed-off-by: Wojciech Malikowski <wojciech.malikowski@intel.com>
Change-Id: I904c2b7fb92bf93ecbe3a036bde976d84e4fb71e
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/qemu/+/2276
Tested-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>