hw/block: check aer_timer is valid when sending events
This is a workaround for issues when the controller isn't fully
initialized (i.e. aer_timer == NULL), but receives a doorbell write that
causes it to send an "invalid doorbell" asynchronous event.  This patch
makes sure that the event is only scheduled if aer_timer is initialized
to avoid the segfault.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I87ea3b15a591f5635fffefce1f341e3d76181aac
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/qemu/+/6969
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Tested-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>